[nxp toup][config][freertos][next gen] Updates to support factory data generation with CMake/Kconfig
Updates in Kconfig definition to use "Macro value is in quotes" hint, which allow the Kconfig to generate macro with quotes without having to use the backslashes.
This is required to ensure "generate_factory_data.cmake" from zephyr can also be used by freertos platforms to generate factory data binaries as a post-build process.

Signed-off-by: Dina Benamar <dina.benamarelmaaroufi@nxp.com>