[nxp toup][cmake][examples][mcxw7x] Fix CONFIG_CHIP_APP_QUEUE_TICKS_TO_WAIT not being propagated to the application
Although CONFIG_CHIP_APP_QUEUE_TICKS_TO_WAIT setting was correctly
defined by the Kcofig, it was not propagated to the
"AppTaskFreeRTOS.cpp" source file because this source file is
not #including the "AppConfig.h" header file. To minimize impact on
the release candidate, we chose to add this to the compiler command
line.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>
Reviewed-by: Dina Benamar El Maaroufi <dina.benamarelmaaroufi@nxp.com>