[nxp toup][cmake][examples][mcxw7x] Fix default systicks callback functions not being replaced with the proper low-power implementation
The default weak callback functions which are updating the FreeRTOS
systicks were not properly replaced with their strong implementations
because both the weak and strong symbols were hosted in the same static
library and this seems to be a linker limitation.

One way to work around this is to make the source files hosting the
implementations of the callback functions part of the application's
source file list so that they appear directly on the application link
command line.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>