[nxp toup][cmake][examples] Allow the user build config file take precedence over the default application build config
The user was unable to change the default application build settings
because the build config file received via the CONF_FILE_NAME CMake
variable was incapable of overwriting the default application build
settings. This update reverses the order of assembly of build settings
and allows the user build config file to take precedence over the
default application build config.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>