[nxp fromtree] [NXP][CMake][MCXW7x] Fix low-power app flavors not operating in low-power mode (#39837)
* [NXP][cmake] Updates to align GN and CMake builds for MCXW platforms

Signed-off-by: Dina Benamar <dina.benamarelmaaroufi@nxp.com>
Reviewed-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][platform][mcxw71] Fix CHIP_DEVICE_CONFIG_ENABLE_EXTENDED_DISCOVERY in CMake build

Signed-off-by: Dina Benamar <dina.benamarelmaaroufi@nxp.com>
Reviewed-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][examples][mcxw7x] Fix low-power app flavors not operating in low-power mode

The low-power build configurations of the Matter apps for MCXW7x
platforms were not operating in low-power mode due to the fact that
an incorrect configuration of build flags was being used.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][examples][common] Fix incorrect header include guards in ButtonsBle.cpp

The "UserInterfaceFeedback.h" header is only required to be included
when UI feedback is enabled. Otherwise it breaks the CMake build for
app flavors that are not using UI feedback.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][examples][common] Fix APP_QUEUE_TICKS_TO_WAIT too low for low-power applications

The APP_QUEUE_TICKS_TO_WAIT configuration value is dictating how
frequent the main Matter application thread is querying the event
queues. For low-power build configurations this frequency needs to be
low enough in order for the application to be able to reach low-power
states.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

---------

Signed-off-by: Dina Benamar <dina.benamarelmaaroufi@nxp.com>
Signed-off-by: Marian Chereji <marian.chereji@nxp.com>
Co-authored-by: dinabenamar <108664279+dinabenamar@users.noreply.github.com>
(cherry picked from commit c659cf9)