MATTER-3487 Fix bug: chip-tool-ele failed to trigger on i.MX9
For chip-tool-ele application, PersistentStorageOperationalKeystoreEle as
a public class will be declare many times in early stage. Also, the
constructor repeatedly invoked.
For i.MX9, we override the constructor to establishes a communication
channel with ele firmware. that will result system hang.

To fix this bug, this patch Abstract the ELE interface into a singleton
class: EleManagerImpl, and use a smart pointer to manage it. This ensures
that only one instance of the ELE interface exists and that resources are
properly released by promptly calling the destructor.

Change-Id: I1bc0e833d920ef009111ce8ea217182b4048fadc
Signed-off-by: Maximus <maximus.sun@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/29383
Reviewed-by: Ji Luo <ji.luo@nxp.com>